Move HyFD's preprocessor logic to separate file
To make it reusable since HyUCC has exactly the same module
